remover pessoa apos 7 dias se o email nao foi confirmado

+++ main.py
@@ -1,66 +1,73 @@
-import smtplib  # Biblioteca para enviar e-mails usando o protocolo SMTP
-from email.mime.multipart import MIMEMultipart  # Classe para criar e-mails com múltiplas partes (texto, anexos, etc.)
-from email.mime.text import MIMEText  # Classe para criar objetos de texto para o e-mail
-import pandas as pd  # Biblioteca para manipulação de dados, especialmente arquivos do Excel
-import time  # Biblioteca para manipulação de tempo (pausas, etc.)
-import schedule  # Biblioteca para agendamento de tarefas
-import uuid  # Biblioteca para gerar identificadores únicos
-from flask import Flask, request, jsonify  # Biblioteca para criar um servidor web
+import smtplib
+from email.mime.multipart import MIMEMultipart
+from email.mime.text import MIMEText
+import pandas as pd
+import time
+import schedule
+import uuid
+import os
+from datetime import datetime, timedelta
+from dotenv import load_dotenv
+from flask import Flask, request, jsonify
+
+# Carregar variáveis de ambiente do arquivo .env
+load_dotenv()
+
+app = Flask(__name__)
+
 
 # Função para enviar e-mail
 def enviar_email(subject, body_html, to_email, from_email, password, smtp_server, smtp_port, from_name):
     try:
-        # Inicia uma conexão com o servidor SMTP usando o contexto "with" para garantir que a conexão seja fechada corretamente
         with smtplib.SMTP(smtp_server, smtp_port) as server:
-            server.starttls()  # Inicia a conexão TLS (Transport Layer Security) para segurança
-            server.login(from_email, password)  # Faz login no servidor SMTP com as credenciais fornecidas
+            server.starttls()
+            server.login(from_email, password)
 
-            # Cria uma mensagem de e-mail com múltiplas partes
             msg = MIMEMultipart()
-            msg['From'] = f"{from_name} <{from_email}>"  # Define o remetente do e-mail
-            msg['To'] = to_email  # Define o destinatário do e-mail
-            msg['Subject'] = subject  # Define o assunto do e-mail
-            msg.attach(MIMEText(body_html, 'html'))  # Anexa o corpo do e-mail no formato HTML
+            msg['From'] = f"{from_name} <{from_email}>"
+            msg['To'] = to_email
+            msg['Subject'] = subject
+            msg.attach(MIMEText(body_html, 'html'))
 
-            server.sendmail(from_email, to_email, msg.as_string())  # Envia o e-mail
-            print(f"E-mail enviado para {to_email}")  # Imprime uma mensagem de confirmação
+            server.sendmail(from_email, to_email, msg.as_string())
+            print(f"E-mail enviado para {to_email}")
     except Exception as e:
-        print(f"Erro ao enviar o e-mail para {to_email}: {e}")  # Em caso de erro, imprime a mensagem de erro
+        print(f"Erro ao enviar o e-mail para {to_email}: {e}")
+
 
 # Função para ler um arquivo HTML
 def ler_html(file_path):
     with open(file_path, 'r', encoding='utf-8') as file:
-        return file.read()  # Lê e retorna o conteúdo do arquivo HTML
+        return file.read()
+
 
 # Função para enviar e-mails iniciais
 def enviar_emails_iniciais():
-    from_email = "daniel@redeabrigo.org"  # E-mail do remetente
-    from_name = "Daniel Mendes"  # Nome do remetente
-    password = "XXXXXXXXX"  # Senha do e-mail do remetente
-    smtp_server = "smtp.gmail.com"  # Servidor SMTP do Gmail
-    smtp_port = 587  # Porta do servidor SMTP (587 para TLS)
+    from_email = os.getenv("EMAIL")
+    from_name = os.getenv("NAME")
+    password = os.getenv("EMAIL_PASSWORD")
+    smtp_server = "smtp.gmail.com"
+    smtp_port = 587
 
-    html_inicial = ler_html('modeloHtml/email1.html')  # Lê o conteúdo do arquivo HTML inicial
+    html_inicial = ler_html('modeloHtml/email1.html')
 
-    # Lê a tabela do Excel com os dados dos e-mails
-    emails_df = pd.read_excel('emails.xlsx', dtype={'Email': str, 'Nome': str, 'Assunto': str, 'Corpo': str, 'Recebido': str, 'ConfirmationID': str})
+    emails_df = pd.read_excel('emails.xlsx',
+                              dtype={'Email': str, 'Nome': str, 'Assunto': str, 'Corpo': str, 'Recebido': str,
+                                     'ConfirmationID': str, 'DataEnvioInicial': str})
 
-    # Itera sobre cada linha do DataFrame
     for index, row in emails_df.iterrows():
-        to_email = row['Email']  # Endereço de e-mail do destinatário
-        subject = row['Assunto']  # Assunto do e-mail
-        body = row['Corpo']  # Corpo do e-mail
-        recebido = row['Recebido']  # Status de recebimento
-        nome = row.get('Nome', 'Prezado(a)')  # Nome do destinatário (ou 'Prezado(a)' se não especificado)
+        to_email = row['Email']
+        subject = row['Assunto']
+        body = row['Corpo']
+        recebido = row['Recebido']
+        nome = row.get('Nome', 'Prezado(a)')
 
-        if pd.isna(to_email):  # Verifica se o endereço de e-mail está ausente (NaN)
-            continue  # Pula para a próxima iteração
+        if pd.isna(to_email):
+            continue
 
-        # Gera um identificador único para o link de confirmação
         confirmation_id = str(uuid.uuid4())
         confirmation_link = f"http://localhost:5000/confirm?email={to_email}&id={confirmation_id}"
 
-        # Variáveis para substituir no HTML do e-mail
         variables = {
             'Nome': nome,
             'nome_da_crianca': 'Maria',
@@ -69,50 +76,50 @@ def enviar_emails_iniciais():
             'confirmation_link': confirmation_link
         }
 
-        # Verifica se o e-mail já foi recebido
         if pd.isna(recebido) or recebido.lower() != 'sim':
             body_html = html_inicial
-            # Substitui placeholders com os valores reais
             for key, value in variables.items():
                 body_html = body_html.replace(f'{{{key}}}', value)
 
-            # Envia o e-mail
             enviar_email(subject, body_html, to_email, from_email, password, smtp_server, smtp_port, from_name)
-            # Armazena o identificador único no DataFrame
             emails_df.at[index, 'ConfirmationID'] = confirmation_id
+            emails_df.at[index, 'DataEnvioInicial'] = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
         else:
             print(f"E-mail já recebido por {to_email}")
 
-    emails_df.to_excel('emails.xlsx', index=False)  # Salva as alterações no arquivo Excel
+    emails_df.to_excel('emails.xlsx', index=False)
 
-# Função para enviar e-mails de reenvio
+
+# Função para enviar e-mails de reenvio e remover campos Email e Nome após 7 dias
 def enviar_emails_reenvio():
-    from_email = "daniel@redeabrigo.org"  # E-mail do remetente
-    from_name = "Daniel Mendes"  # Nome do remetente
-    password = "XXXXXXXX"  # Senha do e-mail do remetente
-    smtp_server = "smtp.gmail.com"  # Servidor SMTP do Gmail
-    smtp_port = 587  # Porta do servidor SMTP (587 para TLS)
+    from_email = os.getenv("EMAIL")
+    from_name = os.getenv("NAME")
+    password = os.getenv("EMAIL_PASSWORD")
+    smtp_server = "smtp.gmail.com"
+    smtp_port = 587
 
-    html_reenvio = ler_html('modeloHtml/email2.html')  # Lê o conteúdo do arquivo HTML de reenvio
+    html_reenvio = ler_html('modeloHtml/email2.html')
 
-    # Lê a tabela do Excel com os dados dos e-mails
-    emails_df = pd.read_excel('emails.xlsx', dtype={'Email': str, 'Nome': str, 'Assunto': str, 'Corpo': str, 'Recebido': str, 'ConfirmationID': str})
+    emails_df = pd.read_excel('emails.xlsx',
+                              dtype={'Email': str, 'Nome': str, 'Assunto': str, 'Corpo': str, 'Recebido': str,
+                                     'ConfirmationID': str, 'DataEnvioInicial': str})
 
-    # Itera sobre cada linha do DataFrame
+    today = datetime.now()
+
     for index, row in emails_df.iterrows():
-        to_email = row['Email']  # Endereço de e-mail do destinatário
-        subject = "Lembrete: Confirmação de Recebimento"  # Assunto do e-mail de reenvio
-        body = row['Corpo']  # Corpo do e-mail
-        recebido = row['Recebido']  # Status de recebimento
-        nome = row.get('Nome', 'Prezado(a)')  # Nome do destinatário (ou 'Prezado(a)' se não especificado)
+        to_email = row['Email']
+        subject = "Lembrete: Confirmação de Recebimento"
+        body = row['Corpo']
+        recebido = row['Recebido']
+        nome = row.get('Nome', 'Prezado(a)')
+        data_envio_inicial = row.get('DataEnvioInicial')
 
-        if pd.isna(to_email):  # Verifica se o endereço de e-mail está ausente (NaN)
-            continue  # Pula para a próxima iteração
+        if pd.isna(to_email):
+            continue
 
-        confirmation_id = row['ConfirmationID']  # Usa o mesmo ID gerado anteriormente
+        confirmation_id = row['ConfirmationID']
         confirmation_link = f"http://localhost:5000/confirm?email={to_email}&id={confirmation_id}"
 
-        # Variáveis para substituir no HTML do e-mail
         variables = {
             'Nome': nome,
             'nome_da_crianca': 'Maria',
@@ -121,28 +128,63 @@ def enviar_emails_reenvio():
             'confirmation_link': confirmation_link
         }
 
-        # Verifica se o e-mail já foi recebido
         if pd.isna(recebido) or recebido.lower() != 'sim':
             body_html = html_reenvio
-            # Substitui placeholders com os valores reais
             for key, value in variables.items():
                 body_html = body_html.replace(f'{{{key}}}', value)
 
-            # Envia o e-mail
             enviar_email(subject, body_html, to_email, from_email, password, smtp_server, smtp_port, from_name)
         else:
             print(f"E-mail já recebido por {to_email}")
 
-    emails_df.to_excel('emails.xlsx', index=False)  # Salva as alterações no arquivo Excel
+        if data_envio_inicial and isinstance(data_envio_inicial, str):
+            data_envio_inicial_dt = datetime.strptime(data_envio_inicial, '%Y-%m-%d %H:%M:%S')
+            elapsed_time = today - data_envio_inicial_dt
+
+            if elapsed_time > timedelta(days=7):
+                print(f"Removendo o email e o nome de {to_email} da lista de e-mails")
+                emails_df.at[index, 'Email'] = ''
+                emails_df.at[index, 'Nome'] = ''
+                emails_df.at[index, 'ConfirmationID'] = ''
+                emails_df.at[index, 'DataEnvioInicial'] = ''
+
+    emails_df.to_excel('emails.xlsx', index=False)
+
+
+@app.route('/confirm', methods=['GET'])
+def confirm_receipt():
+    email = request.args.get('email')
+    confirmation_id = request.args.get('id')
+
+    emails_df = pd.read_excel('emails.xlsx',
+                              dtype={'Email': str, 'Nome': str, 'Assunto': str, 'Corpo': str, 'Recebido': str,
+                                     'ConfirmationID': str, 'DataEnvioInicial': str})
+
+    for index, row in emails_df.iterrows():
+        if row['Email'] == email and row['ConfirmationID'] == confirmation_id:
+            emails_df.at[index, 'Recebido'] = 'sim'
+            emails_df.to_excel('emails.xlsx', index=False)
+            return jsonify({"message": "Confirmação recebida com sucesso!"}), 200
+
+    return jsonify({"message": "Confirmação inválida!"}), 400
+
 
 # Função principal que configura o agendamento
 def main():
-    # Agendar a função enviar_emails_reenvio para rodar a cada 10 dias
-    schedule.every(10).days.do(enviar_emails_reenvio)
-    enviar_emails_iniciais()  # Executa a função enviar_emails_iniciais imediatamente
+    schedule.every(7).days.do(enviar_emails_reenvio)  # Agendar o reenvio a cada 7 dias
+    print("Agendamento iniciado")
+    enviar_emails_iniciais()
     while True:
-        schedule.run_pending()  # Executa as tarefas agendadas se houver
-        time.sleep(1)  # Aguarda 1 segundo antes de verificar novamente
+        schedule.run_pending()
+        time.sleep(1)
+
 
 if __name__ == "__main__":
-    main()  # Executa a função principal se o script for executado diretamente
+    from threading import Thread
+
+    # Iniciar o servidor Flask em uma thread separada
+    flask_thread = Thread(target=lambda: app.run(host='0.0.0.0', port=5000))
+    flask_thread.start()
+
+    # Executar a função principal de agendamento
+    main()
